[aptitude @ Explicitly cast "line" to an unsigned int to avoid build failures on platforms where not doing this tosses a warning.]

+++ src/generic/util/eassert.cc
@@ -34,8 +34,10 @@ std::string AssertionFailure::errmsg() const
 {
   if(msg.empty())
     return ssprintf("%s:%d: %s: Assertion \"%s\" failed.",
-		    file.c_str(), line, func.c_str(), exp.c_str());
+		    file.c_str(), static_cast<unsigned int>(line),
+		    func.c_str(), exp.c_str());
   else
     return ssprintf("%s:%d: %s: %s: Assertion \"%s\" failed.",
-		    file.c_str(), line, func.c_str(), msg.c_str(), exp.c_str());
+		    file.c_str(), static_cast<unsigned int>(line),
+		    func.c_str(), msg.c_str(), exp.c_str());
 }
